requirements.txt
anthropic==0.28.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.0
spacy==3.7.5
wikipedia==1.4.0
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
joblib==1.4.2
streamlit==1.35.0
requests==2.32.2
